pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fairly"
version = "0.1.0"
description = "Bias evaluation toolkit for Vision-Language Models (VLMs)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Nicolás Alayo" },
    { name = "Aleksandra Laskowska" },
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "boto3>=1.34.0",
    "pandas>=2.2.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "fpdf2>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
fairly = "fairly.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/fairly"]

[tool.ruff]
line-length = 88
target-version = "py310"
